+++ brewtils/decorators.py
@@ -40,6 +40,7 @@ def client(
     bg_version=None,  # type: Optional[str]
     group=None,  # type: str
     groups=[],  # type: Optional[List[str]]
+    prefix_topic=None,  # type: Optional[str]
 ):
     # type: (...) -> Type
     """Class decorator that marks a class as a beer-garden Client
@@ -68,6 +69,7 @@ def client(
         bg_version: Optional plugin version
         group: Optional plugin group
         groups: Optional plugin groups
+        prefix_topic: Optional prefix for Generated Command to Topic mappings
 
     Returns:
         The decorated class
@@ -75,7 +77,12 @@ def client(
     """
     if _wrapped is None:
         return functools.partial(
-            client, bg_name=bg_name, bg_version=bg_version, groups=groups, group=group
+            client,
+            bg_name=bg_name,
+            bg_version=bg_version,
+            groups=groups,
+            group=group,
+            prefix_topic=prefix_topic,
         )  # noqa
 
     # Assign these here so linters don't complain
@@ -84,6 +91,7 @@ def client(
     _wrapped._bg_commands = []
     _wrapped._current_request = None
     _wrapped._groups = groups
+    _wrapped._prefix_topic = prefix_topic
 
     if group:
         _wrapped._groups.append(group)