+++ brewtils/schemas.py
@@ -256,6 +256,7 @@ class SystemSchema(BaseSchema):
     local = fields.Bool(allow_none=True)
     template = fields.Str(allow_none=True)
     groups = fields.List(fields.Str(), allow_none=True)
+    prefix_topic = fields.Str(allow_none=True)
 
 
 class SystemDomainIdentifierSchema(BaseSchema):
@@ -633,6 +634,7 @@ class SubscriberSchema(BaseSchema):
     version = fields.Str(allow_none=True)
     instance = fields.Str(allow_none=True)
     command = fields.Str(allow_none=True)
+    subscriber_type = fields.Str(allow_none=True)
 
 
 class TopicSchema(BaseSchema):