+++ CHANGELOG.rst
@@ -6,6 +6,9 @@ Brewtils Changelog
 TBD
 
 - Update Parameter decorator to map parameter type from literal to string value if possible
+- Added Subscriber Types to Subscriber model
+- Added Prefix Topics to System model
+- Support adding Prefix Topics for the Generated Subscribers. It is supported through the `@client` or SystemClient inputs or beer.conf
 
 3.26.2
 ------